refactor: profileView Refactoring User to Myprofile
